Moved `Extension` to `Base.Chord`
