+++ src/Parser.hs
@@ -5,9 +5,10 @@ module Parser
   ) where
 
 import Base.Core.Accidental
-import Base.Core.Extension
 import Base.Core.Quality
 import Base.Chord
+
+import Base.Chord.Extension
 
 import Common.Utils (rightToMaybe)
